Support Swift and Swift for TensorFlow.
To allow the Notebook to separately support both the Swift 4.1 release
version, and the Swift for TensorFlow development version, I've added
two kernel specs to the Jupyter notebook.

Each kernel spec is a JSON file which specifies the command line
arguments to pass to the iSwift executable. I have also updated the
iSwift code so that it will read these command line arguments and use
them to determine the command and command line arguments to use in
order to launch the REPL process.

The iSwift executable does not provide Swift language support. It
merely acts as an adapter between the Jupyter process and the swift or
swift-tensorflow executable.